backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.11
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
celery==5.4.0
redis==5.1.1
httpx==0.27.2
pillow==10.4.0
aiofiles==24.1.0
anthropic>=0.40.0
openai>=1.50.0
google-generativeai>=0.8.0
sentry-sdk[fastapi]>=2.0.0
